+++ .gitignore
@@ -1,2 +1,3 @@
 *.token
 backup
+__pycache__
