+++ fonctions.py
@@ -1,0 +1,86 @@
+from import_et_variable import *
+#--|--# fonctions
+
+
+#fonction pour restrindre certaines commande uniquement au admins
+def admin_restrict(ctx) :
+    if ctx.message.author.id in admin_id_user :
+        return True
+    else :
+        return False
+    
+#fonction pour ajouter une carte. A renseigner : nom, rarete
+def ajouter_une_carte(nom, rarete) :
+    try :
+        baseDeDonnees = sqlite3.connect(f'./assets/database/{db_used}')
+        curseur = baseDeDonnees.cursor()
+        curseur.execute("INSERT INTO Cartes (nom, rarete) VALUES (?, ?)", (f"{nom}", f"{rarete}"))
+        baseDeDonnees.commit()
+        baseDeDonnees.close()
+        return None
+    except Exception as e :
+        return e
+
+
+#fonction qui permet de voir si on a changé de jour. (elle sera présente devant chaque actions) (elle update tout les joueurs en fonction du nombre e jour passé)
+#PROBLEME = Si le bot est inactif ou crash au moment de changer de jour jusqu'à la prochaine action, le changement ne prendra pas effet
+def test_changement_de_jour() :
+    global DATE_actuel
+    jour_diff = int((date.today()-DATE_actuel).days)
+    if int(jour_diff) > 0 :
+        DATE_actuel = date.today()  #date du jour
+        baseDeDonnees = sqlite3.connect(f'./assets/database/{db_used}')
+        curseur = baseDeDonnees.cursor()
+        curseur.execute(f"""UPDATE Joueur 
+                    SET fragment = fragment + {jour_diff}, fragment_cumule = 0, xp = xp + {jour_diff}""")
+        baseDeDonnees.commit()
+        baseDeDonnees.close()
+
+
+#fonction pour tester si un utilisateur est ans la BDD (test seulement)
+def test_player_in_bdd(id_user) :
+    baseDeDonnees = sqlite3.connect(f'./assets/database/{db_used}')
+    curseur = baseDeDonnees.cursor()
+    curseur.execute("SELECT id_discord_player FROM Joueur")
+    result = curseur.fetchall()
+    baseDeDonnees.close()
+    members_in_dbb = [result[k][0] for k in range(len(result))]
+    if id_user in members_in_dbb :
+        return True
+    else :
+        return False
+
+
+#fonction pour tester si un utilisatzur est dans la BDD. Si oui, on lui rajoute ddans la bdd les info nécessaire
+def test_cration_bdd_user(id_user) :
+    if test_player_in_bdd(id_user) == False:
+        baseDeDonnees = sqlite3.connect(f'./assets/database/{db_used}')
+        curseur = baseDeDonnees.cursor()
+        curseur.execute(f"SELECT id FROM Cartes ")
+        result = curseur.fetchall()
+        id_cartes = [result[k][0] for k in range(len(result))]
+        #ensuite, on regarde tout les gens sur le serveur et si un joueur n'es pas dans la BDD, il y est ajouté avec ses stats à 0
+        curseur.execute("INSERT INTO Joueur (id_discord_player, fragment, fragment_cumule, xp, curseur_carte) VALUES (?, ?, ?, ?, ?)", (id_user, 0, 0, 0, 0))
+        for id in id_cartes :
+            curseur.execute("INSERT INTO carte_possede (id_discord_player, id, nombre_carte_possede) VALUES (?, ?, ?)", (id_user, id, 0))    
+        baseDeDonnees.commit()
+        baseDeDonnees.close()
+
+
+#fonction pour formater le nom de l'image en nom utilisate en tant qu'url pour afficher l'image ans un embed
+def formatage_nom_carte(nom) :
+    nom = nom.replace(" ", "_")
+    nom = nom.replace("é", "e")
+    nom = nom.replace("è", "e")
+    nom = nom.replace("ê", "e")
+    nom = nom.replace("à", "a")
+    nom = nom.replace("'", "")
+    return nom
+
+
+#test si un joueur écrit une commande pour ne pa lui ajouter de fragment auquel cas
+def test_joueur_ecrit_commande(msg) :
+    for com in liste_comandes :
+        if com in msg :
+            return True
+    return False
